+++ handler/throttle/config_internal_test.go
@@ -1,4 +1,4 @@
-package ratelimit
+package throttle
 
 import (
 	"context"
@@ -9,7 +9,7 @@ import (
 
 func TestConfig_Errors(t *testing.T) {
 	type testCase struct {
-		configured config.RateLimits
+		configured config.Throttles
 		expMessage string
 	}
 
@@ -25,62 +25,62 @@ func TestConfig_Errors(t *testing.T) {
 
 	for _, tc := range []testCase{
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{PerPeriod: num, PeriodWindow: ""},
 			},
 			"'period' must not be 0 (zero)",
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: min, PeriodWindow: ""},
 			},
 			"'per_period' must not be 0 (zero)",
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: foo, PerPeriod: num, PeriodWindow: ""},
 			},
 			`period: time: invalid duration "foo"`,
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: neg, PerPeriod: num, PeriodWindow: ""},
 			},
 			`period: cannot be negative: '-1s'`,
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: zeroStr, PerPeriod: num, PeriodWindow: ""},
 			},
 			`'period' must not be 0 (zero)`,
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: min, PerPeriod: zeroInt, PeriodWindow: ""},
 			},
 			`'per_period' must not be 0 (zero)`,
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: min, PerPeriod: num, PeriodWindow: ""},
 				{Period: sec, PerPeriod: num, PeriodWindow: ""},
 			},
 			`duplicate period ("60s") found`,
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: min, PerPeriod: num, PeriodWindow: "test"},
 			},
 			`unsupported 'period_window' ("test") given`,
 		},
 		{
-			[]*config.RateLimit{
+			[]*config.Throttle{
 				{Period: min, PerPeriod: num, PeriodWindow: "", Mode: "test"},
 			},
 			`unsupported 'mode' ("test") given`,
 		},
 	} {
-		_, err := ConfigureRateLimits(context.TODO(), tc.configured, nil)
+		_, err := ConfigureThrottles(context.TODO(), tc.configured, nil)
 		if err == nil {
 			t.Fatal("Missing error")
 		}